requirements.txt
djangorestframework-jwt==1.11.0
Django==3.2.16
django-cors-headers==3.7.0
idna==3.4
PyMySQL==1.0.2
PyJWT==1.7.1
python-dateutil==2.6.1
pytz==2023.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
urllib3==1.26.12
djangorestframework==3.12.1
requests==2.28.1
mysqlclient==2.2.0
boto3==1.6.19